project/requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile project/requirements/base.in
#
-e git+https://github.com/darklow/django-suit#egg=django-suit  # via -r project/requirements/base.in
amqp==5.0.1               # via kombu
asgiref==3.2.10           # via django
billiard==3.6.3.0         # via celery
blessings==1.7            # via curtsies
bpython==0.20             # via -r project/requirements/base.in
celery==5.0.1             # via -r project/requirements/base.in
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click-didyoumean==0.0.3   # via celery
click-repl==0.1.6         # via celery
click==7.1.2              # via celery, click-didyoumean, click-repl, flex
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
curtsies==0.3.4           # via bpython
defusedxml==0.6.0         # via odfpy
diff-match-patch==20200713  # via django-import-export
django-appconf==1.0.4     # via django-image-cropping
django-ckeditor==6.0.0    # via -r project/requirements/base.in
django-colorfield==0.3.2  # via -r project/requirements/base.in
django-cors-headers==3.5.0  # via -r project/requirements/base.in
django-image-cropping==1.5.0  # via -r project/requirements/base.in
django-import-export==2.4.0  # via -r project/requirements/base.in
django-js-asset==1.2.2    # via django-ckeditor
django-modeltranslation==0.16  # via -r project/requirements/base.in
django-redis-cache==3.0.0  # via -r project/requirements/base.in
django-solo==1.1.3        # via -r project/requirements/base.in
django==3.1.2             # via -r project/requirements/base.in, django-appconf, django-cors-headers, django-import-export, django-modeltranslation, djangorestframework, drf-yasg, easy-thumbnails
djangorestframework-jwt==1.11.0  # via -r project/requirements/base.in
djangorestframework==3.12.1  # via -r project/requirements/base.in, drf-yasg
drf-yasg==1.20.0          # via -r project/requirements/base.in
easy-thumbnails==2.7      # via -r project/requirements/base.in
et-xmlfile==1.0.1         # via openpyxl
flake8==3.8.4             # via -r project/requirements/base.in
flex==6.14.1              # via -r project/requirements/base.in
greenlet==0.4.17          # via bpython
hiredis==1.1.0            # via -r project/requirements/base.in
idna==2.10                # via requests
inflection==0.5.1         # via drf-yasg
itypes==1.2.0             # via coreapi
jdcal==1.4.1              # via openpyxl
jinja2==2.11.2            # via -r project/requirements/base.in, coreschema
jsonpointer==2.0          # via flex
kombu==5.0.2              # via celery
markdown==3.3.3           # via -r project/requirements/base.in
markuppy==1.14            # via tablib
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
odfpy==1.4.1              # via tablib
openpyxl==3.0.5           # via tablib
packaging==20.4           # via drf-yasg
pep8==1.7.1               # via -r project/requirements/base.in
pillow==8.0.1             # via -r project/requirements/base.in, easy-thumbnails
prompt-toolkit==3.0.8     # via click-repl
psycopg2-binary==2.8.6    # via -r project/requirements/base.in
pycodestyle==2.6.0        # via flake8
pyflakes==2.2.0           # via flake8
pygments==2.7.2           # via bpython
pyjwt==1.7.1              # via djangorestframework-jwt
pyparsing==2.4.7          # via packaging
python-dateutil==2.8.1    # via -r project/requirements/base.in
pytz==2020.1              # via -r project/requirements/base.in, celery, django
pyyaml==5.3.1             # via flex, tablib
raven==6.10.0             # via -r project/requirements/base.in
redis==3.5.3              # via -r project/requirements/base.in, django-redis-cache
requests==2.24.0          # via bpython, coreapi, flex
rfc3987==1.3.8            # via flex
ruamel.yaml.clib==0.2.2   # via ruamel.yaml
ruamel.yaml==0.16.12      # via drf-yasg
simplejson==3.17.2        # via -r project/requirements/base.in
six==1.15.0               # via blessings, bpython, click-repl, django-modeltranslation, flex, packaging, python-dateutil, transliterate
sqlparse==0.4.1           # via django
strict-rfc3339==0.7       # via flex
tablib[html,ods,xls,xlsx,yaml]==2.0.0  # via django-import-export
transliterate==1.10.2     # via -r project/requirements/base.in
uritemplate==3.0.1        # via coreapi, drf-yasg
urllib3==1.25.11          # via requests
uwsgi==2.0.19.1           # via -r project/requirements/base.in
validate-email==1.3       # via flex
vine==5.0.0               # via amqp, celery
wcwidth==0.2.5            # via bpython, curtsies, prompt-toolkit
xlrd==1.2.0               # via tablib
xlwt==1.3.0               # via tablib
